refactor(wallet): import additional wallets
