+++ src/pages/SettingsScreen/WalletListPage.tsx
@@ -1,18 +1,30 @@
-import React, { useContext } from 'react';
-import { ScrollView } from 'react-native';
+import React, { useContext, useLayoutEffect } from 'react';
+import { Pressable, ScrollView } from 'react-native';
 import { AppContext } from 'context/AppContext';
 import { SafeAreaPage } from 'templates/SafeAreaPage';
 import { SettingsStackProps } from 'pages/MainScreen/SettingsPage';
 import { NativeStackScreenProps } from '@react-navigation/native-stack';
 import { NavGroup } from 'components/NavGroup/NavGroup';
-import { NavItem } from 'components/NavGroup/NavItem';
 import { globalStyles } from 'global.styles';
 import { WalletListItem } from './components/MyWallets/WalletListItem';
+import AddBoxIcon from 'assets/add-box-icon.svg';
+import { DarkTheme } from '@react-navigation/native';
 
 type Props = NativeStackScreenProps<SettingsStackProps, 'settings.wallets'>;
 
 export const WalletListPage: React.FC<Props> = ({ navigation }) => {
   const { accountList } = useContext(AppContext);
+
+  useLayoutEffect(() => {
+    navigation.setOptions({
+      headerRight: () => (
+        <Pressable onPress={() => navigation.navigate('settings.create')}>
+          <AddBoxIcon fill={DarkTheme.colors.primary} />
+        </Pressable>
+      ),
+    });
+  }, [navigation]);
+
   return (
     <SafeAreaPage>
       <ScrollView style={globalStyles.page}>
@@ -24,12 +36,6 @@ export const WalletListPage: React.FC<Props> = ({ navigation }) => {
               onPress={() => navigation.navigate('settings.wallet', { index })}
             />
           ))}
-        </NavGroup>
-        <NavGroup>
-          <NavItem
-            title="Add another wallet"
-            onPress={() => navigation.navigate('settings.create')}
-          />
         </NavGroup>
       </ScrollView>
     </SafeAreaPage>
